ViewModel and Fragment changes
