+++ app/src/main/java/org/oppia/app/player/audio/AudioViewModel.kt
@@ -1,11 +1,8 @@
 package org.oppia.app.player.audio
 
-import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.ViewModel
 import org.oppia.app.fragment.FragmentScope
 import javax.inject.Inject
-
-private const val TAG_DIALOG = "LANGUAGE_DIALOG"
 
 /** [ViewModel] for user app usage history. */
 @FragmentScope
@@ -24,25 +21,5 @@ class AudioViewModel @Inject constructor(
 
   private fun getLanguageCode(): String {
     return currentLanguageCode
-  }
-
-  fun languageSelectionClicked(fragmentManager: FragmentManager) {
-    val previousFragment = fragmentManager.findFragmentByTag(TAG_DIALOG)
-    if (previousFragment != null) {
-      fragmentManager.beginTransaction().remove(previousFragment).commitNow()
-    }
-    val dialogFragment = LanguageDialogFragment.newInstance(
-      getDummyAudioLanguageList(),
-      "en"
-    )
-    dialogFragment.showNow(fragmentManager, TAG_DIALOG)
-  }
-
-  private fun getDummyAudioLanguageList(): List<String> {
-    val languageCodeList = ArrayList<String>()
-    languageCodeList.add("en")
-    languageCodeList.add("hi")
-    languageCodeList.add("hi-en")
-    return languageCodeList
   }
 }
