+++ app/src/main/res/layout/audio_fragment.xml
@@ -1,18 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto">
-
+        xmlns:app="http://schemas.android.com/apk/res-auto">
   <data>
-
     <variable
-      name="fragmentManager"
-      type="androidx.fragment.app.FragmentManager" />
-
+      name="audioFragment"
+      type="org.oppia.app.player.audio.AudioFragment"/>
     <variable
       name="viewModel"
-      type="org.oppia.app.player.audio.AudioViewModel" />
+      type="org.oppia.app.player.audio.AudioViewModel"/>
   </data>
-
   <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
@@ -23,7 +19,6 @@
     android:gravity="center"
     android:orientation="horizontal"
     android:padding="4dp">
-
     <ImageView
       android:id="@+id/ivPlayPauseAudio"
       android:layout_width="wrap_content"
@@ -32,8 +27,7 @@
       android:padding="8dp"
       app:layout_constraintBottom_toBottomOf="parent"
       app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent" />
-
+      app:layout_constraintTop_toTopOf="parent"/>
     <SeekBar
       android:id="@+id/sbAudioProgress"
       style="@style/AudioSeekBar"
@@ -44,21 +38,20 @@
       app:layout_constraintBottom_toBottomOf="parent"
       app:layout_constraintEnd_toStartOf="@+id/tvAudioLanguage"
       app:layout_constraintStart_toEndOf="@+id/ivPlayPauseAudio"
-      app:layout_constraintTop_toTopOf="parent" />
-
+      app:layout_constraintTop_toTopOf="parent"/>
     <TextView
       android:id="@+id/tvAudioLanguage"
       android:layout_width="wrap_content"
       android:layout_height="wrap_content"
+      android:onClick="@{(v) -> audioFragment.languageSelectionClicked()}"
       android:padding="8dp"
+      android:text="@{viewModel.audioLanguageCode}"
       android:textAllCaps="true"
       android:textColor="@color/white"
       android:textSize="14sp"
       android:textStyle="bold"
-      android:text="@{viewModel.audioLanguageCode}"
-      android:onClick="@{(v) -> viewModel.languageSelectionClicked(fragmentManager)}"
       app:layout_constraintBottom_toBottomOf="parent"
       app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintTop_toTopOf="parent" />
+      app:layout_constraintTop_toTopOf="parent"/>
   </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
